Expand AutoUpdateTweaks to ModCollectionHooks
